pyproject.toml
[tool.ruff.lint]
fixable = ["ALL"]
unfixable = ["F401"]

[tool.ruff]
line-length = 120

[tool.poetry]
name = "bioexperiment-suite"
version = "0.1.4"
description = "Python toolbox for managing biological experiment devices including pumps, cell density detectors, thermometers, and other related equipment."
license = "MIT"
authors = ["Denis Khamitov <hamitov.97@mail.ru>"]
readme = "README.md"
repository = "https://github.com/denis240997/bio_tools"
packages = [
    { include = "bioexperiment_suite", from = "src"},
]
include = [
    { path = "src/bioexperiment_suite/device_interfaces.json"},
]

[tool.poetry.dependencies]
python = "^3.12"
loguru = "^0.7.2"
munch = "^4.0.0"
pyserial = "^3.5"
ttkbootstrap = { version = "^1.10.1", optional = true }

[tool.poetry.extras]
gui = ["ttkbootstrap"]

[tool.poetry.scripts]
run_gui = "gui.main:main"

[tool.poetry.group.dev.dependencies]
munch-stubs = "^0.1.2"
ruff = "^0.5.4"
types-pyserial = "^3.5.0.20240527"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs-material = "^9.5.30"
mkdocs = "^1.6.0"
mkdocstrings = {extras = ["python"], version = "^0.25.2"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
